+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/records/RecordTypeConverter.kt
@@ -19,7 +19,6 @@ import expo.modules.kotlin.types.TypeConverterProvider
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
 import kotlin.reflect.KType
-import kotlin.reflect.full.createInstance
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.memberProperties
 import kotlin.reflect.jvm.javaField
@@ -39,8 +38,7 @@ class RecordTypeConverter<T : Record>(
         return@mapNotNull property to PropertyDescriptor(
           typeConverter,
           fieldAnnotation,
-          isRequired = property.findAnnotation<Required>() != null,
-          validators = getValidators(property)
+          isRequired = property.findAnnotation<Required>() != null
         )
       }
       .toMap()
@@ -88,15 +86,6 @@ class RecordTypeConverter<T : Record>(
             descriptor.typeConverter.convert(this, context, forceConversion)
           }
 
-          if (casted != null) {
-            descriptor
-              .validators
-              .forEach { validator ->
-                @Suppress("UNCHECKED_CAST")
-                (validator as FieldValidator<Any>).validate(casted)
-              }
-          }
-
           javaField.isAccessible = true
           javaField.set(instance, casted)
         }
@@ -110,25 +99,9 @@ class RecordTypeConverter<T : Record>(
     return objectConstructorFactory.get(clazz)
   }
 
-  private fun getValidators(property: KProperty1<out Any, *>): List<FieldValidator<*>> {
-    return property
-      .annotations
-      .map findValidators@{ annotation ->
-        val binderAnnotation = annotation.annotationClass.findAnnotation<BindUsing>()
-          ?: return@findValidators null
-        annotation to binderAnnotation
-      }
-      .filterNotNull()
-      .map { (annotation, binderAnnotation) ->
-        val binderInstance = binderAnnotation.binder.createInstance() as ValidationBinder
-        binderInstance.bind(annotation, property.returnType)
-      }
-  }
-
   private data class PropertyDescriptor(
     val typeConverter: TypeConverter<*>,
     val fieldAnnotation: Field,
-    val isRequired: Boolean,
-    val validators: List<FieldValidator<*>>
+    val isRequired: Boolean
   )
 }